requirements.txt
customtkinter
rapidocr-onnxruntime
opencv-python
pillow
mss
shapely
